Переносит константы в константы
