+++ src/const.ts
@@ -23,6 +23,16 @@ const sortTypes: GenericObject = {
   topRated: 'Top rated first',
 } as const;
 
+const ratingTypes = [5, 4, 3, 2, 1] as const;
+
+const ratingTypesMap: GenericObject = {
+  '1': 'Terribly',
+  '2': 'Badly',
+  '3': 'Not bad',
+  '4': 'Good',
+  '5': 'Perfect'
+};
+
 const CITIES: Array<string> = ['Paris', 'Cologne', 'Brussels', 'Amsterdam', 'Hamburg', 'Dusseldorf'];
 
-export {AuthorizationStatus, AppRoute, CITIES, sortTypes, GenericObject};
+export {AuthorizationStatus, AppRoute, CITIES, sortTypes, ratingTypes, ratingTypesMap};
